more stability improvements

+++ test/globus-test-gram-local.pl
@@ -70,65 +70,84 @@ sub test_gram_local
     ($rc, $job_id) = 
         $u->command("globus-job-submit \"$gatekeeper_url\" /bin/echo I ran",5);
 
-    # globus-job-get-output of stdout from gatekeeper_url of jobid $job_id.
-    ($rc, $output) = $u->command("globus-job-get-output -resource " .
-                                 "\"$gatekeeper_url\" $job_id",5);
-    $output =~ /I ran/ && $rc == 0 ? 
-        $u->report("SUCCESS") : $u->report("FAILURE");
-
-    # globus-job-clean on gatekeeper_url of jobid $job_id.
-    ($rc, $output) = $u->command("globus-job-clean -force -resource " .
-                                 "\"$gatekeeper_url\" $job_id",5);
-    $rc == 0 ? $u->report("SUCCESS") : $u->report("FAILURE");
-
-
+    if($rc == 0)
+    {
+        # globus-job-get-output of stdout from gatekeeper_url of jobid $job_id.
+        ($rc, $output) = $u->command("globus-job-get-output -resource " .
+                                     "\"$gatekeeper_url\" $job_id",5);
+        $output =~ /I ran/ && $rc == 0 ? 
+            $u->report("SUCCESS") : $u->report("FAILURE");
+        
+        # globus-job-clean on gatekeeper_url of jobid $job_id.
+        ($rc, $output) = $u->command("globus-job-clean -force -resource " .
+                                     "\"$gatekeeper_url\" $job_id",5);
+        $rc == 0 ? $u->report("SUCCESS") : $u->report("FAILURE");
+    }
+    else
+    {
+        $u->report("FAILURE");
+    }
+        
     # globus-job-submit.  Store jobid in $job_id
     ($rc, $job_id) = 
         $u->command("globus-job-submit \"$gatekeeper_url\" " .
                     "/bin/sh -c \'/bin/echo I ran 1>&2\'",5);
+    if($rc == 0)
+    {   
+        # globus-job-get-output of stderr from gatekeeper_url of jobid $job_id.
+        ($rc, $output) = $u->command("globus-job-get-output -err -resource " .
+                                     "\"$gatekeeper_url\" $job_id",5);
+        $output =~ /I ran/ && $rc == 0 ? 
+            $u->report("SUCCESS") : $u->report("FAILURE");
+        
+        # globus-job-status from gatekeeper_url of jobid $job_id.
+        ($rc, $output) = $u->command("globus-job-status $job_id",5);
+        $output =~ /DONE/ && $rc == 0 ? 
+            $u->report("SUCCESS") : $u->report("FAILURE");
 
-    # globus-job-get-output of stderr from gatekeeper_url of jobid $job_id.
-    ($rc, $output) = $u->command("globus-job-get-output -err -resource " .
-                                 "\"$gatekeeper_url\" $job_id",5);
-    $output =~ /I ran/ && $rc == 0 ? 
-        $u->report("SUCCESS") : $u->report("FAILURE");
 
-    # globus-job-status from gatekeeper_url of jobid $job_id.
-    ($rc, $output) = $u->command("globus-job-status $job_id",5);
-    $output =~ /DONE/ && $rc == 0 ? 
-        $u->report("SUCCESS") : $u->report("FAILURE");
-
-
-    # globus-job-clean on gatekeeper_url of jobid $job_id.
-    ($rc, $output) = $u->command("globus-job-clean -force -resource " .
-                                 "\"$gatekeeper_url\" $job_id",5);
-    $rc == 0 ? $u->report("SUCCESS") : $u->report("FAILURE");
+        # globus-job-clean on gatekeeper_url of jobid $job_id.
+        ($rc, $output) = $u->command("globus-job-clean -force -resource " .
+                                     "\"$gatekeeper_url\" $job_id",5);
+        $rc == 0 ? $u->report("SUCCESS") : $u->report("FAILURE");
+    }
+    else
+    {
+        $u->report("FAILURE");
+    }
  
     # globus-job-submit.  Store jobid in $job_id
     ($rc, $job_id) = 
         $u->command("globus-job-submit \"$gatekeeper_url\" " .
                     "/bin/sh -c \'echo I ran;sleep 120\'",5);
 
-    # globus-job-status from gatekeeper_url of jobid $job_id.
-    ($rc, $output) = $u->command("globus-job-status $job_id",5);
-    $output =~ /ACTIVE/ && $rc == 0 ? 
-        $u->report("SUCCESS") : $u->report("FAILURE");
-    
-    # globus-job-cancel on gatekeeper_url of jobid $job_id.
-    ($rc, $output) = $u->command("globus-job-cancel -force -resource " .
-                                 "\"$gatekeeper_url\" $job_id",5);
-    $rc == 0 ? $u->report("SUCCESS") : $u->report("FAILURE");
-
-    # globus-job-get-output of stdout from gatekeeper_url of jobid $job_id.
-    ($rc, $output) = $u->command("globus-job-get-output -resource " .
-                                 "\"$gatekeeper_url\" $job_id",5);
-    $output =~ /I ran/ && $rc == 0 ? 
-        $u->report("SUCCESS") : $u->report("FAILURE");
-
-    # globus-job-clean on gatekeeper_url of jobid $job_id.
-    ($rc, $output) = $u->command("globus-job-clean -force -resource " .
-                                 "\"$gatekeeper_url\" $job_id",5);
-    $rc == 0 ? $u->report("SUCCESS") : $u->report("FAILURE");
+    if($rc == 0)
+    {
+        # globus-job-status from gatekeeper_url of jobid $job_id.
+        ($rc, $output) = $u->command("globus-job-status $job_id",5);
+        $output =~ /ACTIVE/ && $rc == 0 ? 
+            $u->report("SUCCESS") : $u->report("FAILURE");
+        
+        # globus-job-cancel on gatekeeper_url of jobid $job_id.
+        ($rc, $output) = $u->command("globus-job-cancel -force -resource " .
+                                     "\"$gatekeeper_url\" $job_id",5);
+        $rc == 0 ? $u->report("SUCCESS") : $u->report("FAILURE");
+        
+        # globus-job-get-output of stdout from gatekeeper_url of jobid $job_id.
+        ($rc, $output) = $u->command("globus-job-get-output -resource " .
+                                     "\"$gatekeeper_url\" $job_id",5);
+        $output =~ /I ran/ && $rc == 0 ? 
+            $u->report("SUCCESS") : $u->report("FAILURE");
+        
+        # globus-job-clean on gatekeeper_url of jobid $job_id.
+        ($rc, $output) = $u->command("globus-job-clean -force -resource " .
+                                     "\"$gatekeeper_url\" $job_id",5);
+        $rc == 0 ? $u->report("SUCCESS") : $u->report("FAILURE");
+    }
+    else
+    {
+        $u->report("FAILURE");
+    }
  
     my ($server_pid, $server_fd) = 
         $u->command_blocking("globus-gass-server -c");
@@ -149,16 +168,25 @@ sub test_gram_local
 
     ($rc, $output) = 
         $u->command("globus-url-copy file:/etc/group $server_url$tmpfile",5);
-    $rc == 0 ? $u->report("SUCCESS") : $u->report("FAILURE");
+    
+    if($rc == 0)
+    {
+        $u->report("SUCCESS");
 
-    ($rc, $output) = $u->command("diff /etc/group $tmpfile");
-    $output eq "" && $rc == 0 ? $u->report("SUCCESS") : $u->report("FAILURE");
-
-    $u->command("rm -rf $tmpfile");
-
-    # server-shutdown is stupid, it always returns non zero
-
-    $u->command("globus-gass-server-shutdown $server_url",5);
+        ($rc, $output) = $u->command("diff /etc/group $tmpfile");
+        $output eq "" && $rc == 0 ? 
+            $u->report("SUCCESS") : $u->report("FAILURE");
+        
+        $u->command("rm -rf $tmpfile");
+        
+        # server-shutdown is stupid, it always returns non zero
+        
+        $u->command("globus-gass-server-shutdown $server_url",5);
+    }
+    else
+    {
+        $u->report("FAILURE");
+    }
 
     # the below will kill the server after 5 minutes
